customer_service/requirements.txt
ipykernel
langgraph
langchain-openai
langchain-core
langchain
python-dotenv
streamlit
langchain-community